Update opt-out.md
As per Peter, "There will be no event notification for opt-in incident ."   is an error. The event will be sent to the same subscription as for opt-out. To distinguish optin from optout there will be "isActive" attribute in the event. isActive=true means optout and isActive=false means optin.

So fixed it in documentation now.